set the user header on new requests in IndexAuditTrail
The IndexAuditTrail was not setting the appropriate user header on requests to see if the index exists and
the mapping is updated. This did not fail in tests because we set shield.user, but fails during a normal
installation.

Closes elastic/x-pack#626

Original commit: elastic/x-pack@6825d5fca17c55aec513506c285d48086243c9d6