Don’t use the off line local repository for Maven
Using a shared repository is not recommended and not safe the failing linux build is probably due to this reason as it is complaining about not getting a file lock on the repository. It have just have been luck that is hasn’t happened before.

As before any additional dependencies added to the Java build must be added to NOTICIES.txt


Former-commit-id: 4f472ca7abcc847c864030cfdfe48cc306644ace

Original commit: elastic/x-pack@b063304c6e6889971e642e4cdbb70013fe75f7d6