update settings for tribes to fail if shield is not enabled or mandatory
In 2.0, plugins cannot specify mandatory settings, they can only specify a default additional
set of settings. For tribe nodes, we require shield to be enabled and be a mandatory plugin.
If the settings specified by the user conflict with this, we now throw an exception and fail
startup.

Closes elastic/x-pack#426

Original commit: elastic/x-pack@79a459ceddae53d00d8030e5c47315039ddc772e