create basic interface for a caching realm
This extracts an interface for methods that a caching realm must implement. The idea for this
interface is to allow for custom realms to integrate easily with the clear cache api. With this change
a custom realm can implement this API and the clear cache action can call the evict or evictAll
methods on the realm.

Closes elastic/x-pack#544

Original commit: elastic/x-pack@1ddb0a61d5a2b8795d750d2492bf0a61a8f1cad6