audit tampered request when verifying the user header fails
When we receive a message and cannot decode the signed user header, we should audit
this as a tampered request. The request may not be tampered if nodes are configured
differently in terms of system keys, but this at least provides some evidence of the bad
request on the nodes that receives it.

Closes elastic/x-pack#647

Original commit: elastic/x-pack@e2eac98728bf697403e337cd6e5814b835aa3e8b