test: update certificates to have more localhost subject alternative names
This updates our test certificates to have more subject alternative names and adds a script
to make regeneration easier in the future. Also added are the private keys in PEM format and
a .p12 version of the keystore. In the future we may want to start supporting PKCS12 keystores
and this will be a starting point for that.

Original commit: elastic/x-pack@51edd8d215aba257b3b143565f001047cd486738