Remove shard level injector usage
This commit catches up with master which removed the shard level injector
and changed the logic how the SearcherWrapper works and is installed.
The way we now install it is via a package private onModule(IndexModule) call.
There is no public API for this anymore.
The wrapper also doesn't need to watch the IndexShards state since now it will only
be used when the shard is STARTED or RECOVERED.

Original commit: elastic/x-pack@3b91cbc0bbef75ef8fd05de96bfe05ae92f9c770