Marvel: Fix memory leak when something goes wrong in LocalExporter
When the bulk request executed by the LocalExporter has at least 1 failure, an exception is thrown and the BulkRequestBuilder is not reset. Then the next time Marvel collects data it will contain all the previous requests plus the new ones... leading to an OOM. IThis behavior was reproduced with a closed marvel index but it also happen in case of unassigned shards.

Original commit: elastic/x-pack@a6119bb8ea69fc0a344277f01797faaabd766058