fix logging audit trail to not cause guice issues
When the logging audit trail is configured to add the node hostname or ip address
as a prefix, the logging audit trail can invoke guice dependency injection issues since
the transport that is injected is a proxy. This change makes the logging audit trail a
lifecycle component and waits for the transport to be started before initializing the
prefix.

Closes elastic/x-pack#1104

Original commit: elastic/x-pack@3d229683091519cd42ec34b267e9d8a88929255a