Store Elasticsearch queries in a field called "query"
The "search" object that previously enclosed this was unnecessary.  Also,
in the future we'll want to add an "aggregations" object and it makes
sense for that to go at the same level of nesting as "query".


Former-commit-id: cb039435ce2db9cec640ed6c91a12baeda4d8ce8

Original commit: elastic/x-pack@78a24d5c9e4babefe0a265ed999f2f6a81336196