Replace apache http client with jetty
This also fixes a non-deterministic issue
with the behaviour of streaming data to the engine API.

The issue is that given a thread is writing on the stream
that is being uploaded to the data end-point, if an early
server response is triggered (e.g. too many bad timestamps),
there exists a chance that the writing thread will try to
write into a closed stream. When the server raises an exception,
it closes the stream. It then converts that exception into
a response. However, if in the meantime the thread tries to write
on the stream, the exception is picked up by apache HTTP client
and returned as response instead of the original error.

Using Jetty allows us to solve this by taking advantage of
its deferred content provider mechanism. This should put an
end to the sporadic failures of the integration tests and prevent
such issue from users of the java client.

Finally, note that the Jetty HTTP client  appears to be a cleaner API.
It has allowed simplifying a bit the code in EngineApiClient.


Former-commit-id: b2c6e1448b2aa753f89c49137fe3b75392523b57

Original commit: elastic/x-pack@1aa66485b9f6a69190f48a56ea2a89cbf4eab8cc