EL16-161 Create schedulers for stopped jobs with SchedulerConfig
The reason jobs whose scheduler was stopped cannot be restarted
after the engine is restarted is that no scheduler is created
for them.

This commit is creating a scheduler for all jobs with SchedulerConfig.
The scheduler is started only for jobs whose schedulerStatus is STARTED.


Former-commit-id: ac11fdf7c3843f0d41779cc48592b8698e743dfb

Original commit: elastic/x-pack@7d7e338e5585283fa5ab717cc47cfb133534a3dc