Remove redundant clean up code
Now that the process and timeout entries for a job
always get removed upon closing a job, it is impossible
that an entry exists for the timeouts while no entry exists
for the process map. Thus, the code removing the timeout
entry when no process was found is unnecessary.


Former-commit-id: 8db48d355cce119252e9cb4f896aa3a7500c794b

Original commit: elastic/x-pack@03ce79fcfb82b5ab3e8631131948176811f96fd6