shield: add support for new privilege naming
This commit adds support for the privilege naming defined in elastic/x-pack#1342 and removes the
support for the privileges that were deprecated in 2.3. This change also includes
updates to the documentation to account for the new roles format.

Original commit: elastic/x-pack@c446c92eb20f4f7988ccb45306e156f776c54982