Upon delete, stop scheduler before getting ticket for DELETE
If the scheduler is stopped after obtaining the ticket for the
DELETE action, it is possible that a JobInUseException will
be logged due to the scheduler being in the middle of writing
data.

We can avoid that by simply stopping the scheduler before
acquiring the ticket.

I am also changing the scheduler/jobId map to be a ConcurrentHashMap
as it is being accessed by multiple threads and correct visibility
should be ensured.


Former-commit-id: b4f48d0fefeb31d951ceba1e9409a77f243274a8

Original commit: elastic/x-pack@8a26d860b6ebbb189acd21583647b51bc7fe7292