Adds tombstones to cluster state for index deletions
Previously, we would determine index deletes in the cluster state by
comparing the index metadatas between the current cluster state and the
previous cluster state and decipher which ones were missing (the missing
ones are deleted indices).  This led to confusing logic where it was
difficult to decipher if a missing index in the cluster state was truly
due to a delete and must be deleted on all nodes, or if it was due to
the cluster state being wiped out on master and thus, we want the
indices on the various nodes in the cluster to be imported as dangling
indices.

This commit introduces the notion of index tombstones in the cluster
state, where we are explicit about which indices have been deleted.
Index deletion on each node is now based on the presence of these
tombstones in the cluster state.  There is also functionality to purge
the tombstones after a 14 day expiry, so they don't stick around in the
cluster state forever.

Closes #16358