EL16-285 Change extractor time format to date_time to support 1.x
The query sent to elasticsearch for pulling data specifies a
time range where the format was set to "epoch_millis". It turns
out "epoch_millis" support was added since version 2.0. Thus,
firing the query against a 1.x elasticsearch cluster fails.

This commit changes the format of the time range to date_time
which is the ISO format and is supported both in 1.x and 2.x.


Former-commit-id: 03d8dbc8eeec360c1d7fed179154ba4a0af3ba54

Original commit: elastic/x-pack@1367e2ffb29ddb08138facd8bfa203443be999be