EL16-290 Improve lookback performance
Performing a single scroll search for the entire interval
of the lookback performs much slower compared to chunking
the interval into smaller ones and searching each one of them.

This commit changes the way data extraction is performed when
the interval spans longer than one hour. It calculates the chunk
size so that each search will be expected to return SIZE * shards
documents and iteratively executes searches that span over the chunk
size.

In addition, if a search returns zero hits, it reconfigures the chunk
size and it jumps to the next time when results exist.

This approach appears to speed up lookback searches significantly.
Testing was still ongoing when the commit was pushed but there were
scenarios where the speedup was > 7x.


Former-commit-id: 833be69dc70232be6301119e65f45e89b1264568

Original commit: elastic/x-pack@1ad6902a9cc40e3ae37b9592f76b6c4281faf59a