Increase robustness of client data streaming
In the case where data is sent and the server replies with
an error there is a race condition between the client side
which is still sending data and the server that closes the
connection due to the error response. This seems to still
be surfacing in 8% as described in EL16-300.

This commit does two things:
  - Adds some defensive checks in order to stop sending
  data after a response was received from the server.
  - When the content variable is null, it is converted to
  empty string in order to avoid a NPE reaching the users.
  The error response will still have been lost in this case
  but at least users won't see the NPE.


Former-commit-id: 2c34f6bb03d1e7ec41b9dbbae04bf31129bd4b3d

Original commit: elastic/x-pack@2d727979de949a6cd9d8d538fa0f5081754429d3