Watcher: Fall back on default format color in hipchat action
Our documentation states that we have default attributes for
message.format and message.color, which in fact we do not have
as an NPE was triggered in that case.

This commit falls back to unset defaults and allows for hipchat messages
to be sent without having to configure color/format in the action
or the account.

Closes elastic/x-pack#1666

Original commit: elastic/x-pack@47010b050fe4b38ac96d80c8ee8621c5da5c39bc